getting rid of old files, renamed linkage2bed to merlin2bed
Fixed bug in merlin2bed.pl where a peak that continued through the end of the chromosome didn't get printed to output
Make plink2merlin use both the CompleteMaps and Linkage files for getting genetic map position and remove all variants that are not in one of the two files.  Currently this script assumes that the CompleteMaps and Linkage/freqs/grid* files agree about genetic map position.